xcode 6 iOS 8
converting to xcode 6 iOS 8

+++ TrackMap/TrackMap/MapMenu/mapMenuViewController.m
@@ -48,13 +48,13 @@
 #pragma mark Table View Delegate Methods
 
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (long)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return self.menuItems.count;
 }
 
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath;
+- (double)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath;
 {
     return 60.0;
 }
